feat(auth): 데이터베이스 스키마 변경에 대응하여 Spring Security 기능 수정
- `User` -> `UserData`
- `UserRepository` -> `UserDataRepository`
- #36 참조
