--- src/main/java/moanote/backend/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package moanote.backend.repository;
-
-import moanote.backend.entity.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String username);
-}
