+++ src/main/java/moanote/backend/service/NoteService.java
@@ -1,21 +1,46 @@
 package moanote.backend.service;
 
+import jakarta.transaction.Transactional;
 import moanote.backend.entity.Note;
+import moanote.backend.entity.NoteUserData.Permission;
 import moanote.backend.repository.NoteRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import moanote.backend.repository.NoteUserDataRepository;
+import moanote.backend.repository.UserDataRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.UUID;
 
 @Service
 public class NoteService {
 
-  @Autowired
-  private NoteRepository noteRepository;
+  private final NoteRepository noteRepository;
 
-  public Note createNote() {
-    return noteRepository.createNote();
+  private final UserDataRepository userDataRepository;
+
+  private final NoteUserDataRepository noteUserDataRepository;
+
+  public NoteService(NoteRepository noteRepository, UserDataRepository userDataRepository,
+      NoteUserDataRepository noteUserDataRepository) {
+    this.noteRepository = noteRepository;
+    this.userDataRepository = userDataRepository;
+    this.noteUserDataRepository = noteUserDataRepository;
+  }
+
+  /**
+   * 새로운 Note 를 생성하고, 생성자에게 노트에 대한 OWNER 권한을 부여합니다.
+   *
+   * @param creatorId 생성자 id
+   * @return 생성된 Note entity
+   */
+  @Transactional
+  public Note createNote(UUID creatorId) {
+    Note newNote = noteRepository.createNote();
+    noteUserDataRepository.createNoteUserData(userDataRepository.findById(creatorId).orElseThrow(),
+        newNote,
+        Permission.OWNER);
+    return newNote;
   }
 
   /**
@@ -25,7 +50,7 @@ public class NoteService {
    * @return 찾아진 Note entity 객체
    * @throws NoSuchElementException noteId 에 해당하는 객체를 찾을 수 없는 경우
    */
-  public Note getNoteById(Long noteId) {
+  public Note getNoteById(UUID noteId) {
     return noteRepository.findById(noteId).orElseThrow();
   }
 
@@ -37,11 +62,11 @@ public class NoteService {
    * @return 업데이트 대상이 된 Note entity
    * @throws NoSuchElementException noteId 에 해당하는 객체를 찾을 수 없는 경우
    */
-  public Note updateNote(Long noteId, String updatedContent) {
+  public Note updateNote(UUID noteId, String updatedContent) {
     return noteRepository.updateNote(noteId, updatedContent);
   }
 
-  public List<Note> getNotesByUserId(Long userId) {
-    return noteRepository.findByUserId(userId);
+  public List<Note> getNotesByOwnerUserId(UUID userId) {
+    return noteRepository.findNotesByOwner(userId);
   }
 }
